Add community docs to side panel.
Something that will definitely help make the learning process easier in
the meantime as we work on the larger doc reorganization. Added to the
current side panel. This arguably makes things a little long / crowded.
But I think the benefit > the cost as we try and figure out how to
create a better layout (like a Wiki layout of sorts, as we've
discussed).

The two community docs that come to mind are Joe's and Fang's. Are there
any more?