Arvo section `# Quickstart` to `# Arvo`.
I get that the idea originally here was for this section to be the "get
your hands dirty with Hoon fast" section. But for newcomers, I'd be
shocked if some people didn't confuse this with the place to start
overall. I.e., it doesn't make sense for this to be the Quickstart
section if `Install` is in a different section. Just `# Arvo` is fine.
It also makes the bottom grid row look prettier, as it's then the
progression: Arvo Hoon Nock. Plus, for learning Hoon, we likely want to
direct people to the Urbytes now, as many of us relay in Talk.

The url's are also docs/arvo so this just makes sense.