Reorganize Urbytes and Community docs sections into single 'Learn' section.
What I like about this is that it allows multiple doc series to get
actively developed at the same time in one section that clearly stands
out as the section for contributors to fork and PR. I.e., Curtis has
been busy lately so hasn’t really been able to get to the next Urbyte.
But other docs have been written since then. With a single ‘Learn’
section (I can’t think of a better name — I think the clearness of the
word trumps the potential blandness), people can know where to come
back to if they’ve “graduated” from the existing material and see if
anything new has been added. Basically, it takes less pressure off of
us for managing our own in-house docs and groups them with all the docs
that the community is feeling like writing themselves. I think that’s
how this should be anyways.

@galenwp, @cgyarvin, thoughts?