Temporary Hoon sort order fix.
Exercises are supposed to come after the rest of the Hoon docs, but the
sorting bug is placing them second.

I’m about to fix this bug in the ES6 branch I believe, but we might
just want to get this up now anyways.

Atom also deleted some trailing whitespace.