Arvo network message doc: Boot a comet, not a moon
A number of developers new to Urbit have been reading these docs and
running these examples, only to realize that booting a moon ships their
planets for unknown reasons. This is a simple doc fix that can push
people to comets in the meantime. And this is probably easier to do
than starting a moon anyways.

But maybe all of these docs should point people towards doing these on
fake ships? In an ideal Urbit world, dumb examples never sink your
planet, but unfortunately this can be the case now, and perhaps we
should make that explicit to early developer adopters.