Update examples to use modern hoon. Fixes #180.
(mo is no longer supported. Use (my which is currently considered
the proper way to do this.

Corrects a couple other small syntax errors which were preventing some
of the examples from running correctly. (e.g. incorrect back tics
instead of single quotes and using a / instead of = on line 81)