Adjusting Description of dot-tar
There is a sense in which `.*` is like Nock 2, but I think that's misleading.  `p` and `q` are evaluated against the Hoon subject using Hoon rules, and then the latter product is evaluated against the former using Nock rules.  This isn't really Nock 2, which is, of course, entirely in Nock.  The current description confused me when I was first learning Nock.

This is a nitpick, but I think we can, and should, be a little more clear here.