Removed broken `? ~~Ürbit` example
This example is broken (issue #145).

We still have people trying this out in :dojo and becoming confused when it doesn't work.  Someone was struggling with this today and brought it up in urbit-meta.  We should delete it for now.  Broken examples are frustrating.  When the problem in #145 is corrected, then we can (and should) add it back in.