Normalize Using section, styling, layout; remove /docs/using/urb
The urb repo is outdated, I think, and there's not really any good reason to use it if that's the case. Maybe we can fix it and add it back if it turns out to be something people actually want. But for now, this simplifies things and doesn't point new users to things that might be broken or weird.