Updated egg-timer.md to static Gall
Complete rewrite of egg-timer.hoon example code and walkthrough, to work with static Gall.

Thanks to ~hastuc for walking me through this and ~palfun for helping with style/standards and my own understanding.

Regarding the tutorial itself, I tried to be consistent with other documentation, however I had to make best guesses in some cases. e.g., should we say `++on-poke` or `++  on poke`? Should we capitalize Core, Door, Arm?

Will appreciate review both for style and content.